feat: Modularize state management by extracting data persistence, session management, and todo operations into separate modules
- Implemented TodosPersistence for managing todo storage operations.
- Created TrashPersistence for handling trashed todos.
- Developed SessionPersistence to manage session data.
- Introduced SessionManager to handle session-related operations and state management.
- Added TodoOperations for CRUD operations on todos with server sync capabilities.
- Enhanced error handling and logging throughout the new modules.
